:construction: refactor left pannel
