+++ packages/esm-admin-app/src/components/users/manage-users/manage-user.scss
@@ -22,7 +22,7 @@
 
 .tile {
   border: 1px solid colors.$gray-20;
-  padding: 1.5rem 0;
+  padding: layout.$spacing-05 0;
   text-align: center;
 }
